chore: make history preprocessing reusable, add section
