+++ app/components/Analyst/History/HistoryTable.tsx
@@ -1,9 +1,8 @@
 import styled from 'styled-components';
 import { graphql, useFragment } from 'react-relay';
-import { getFiscalQuarter, getFiscalYear } from 'utils/fiscalFormat';
 import { HistoryTable_query$key } from '__generated__/HistoryTable_query.graphql';
 import { useMemo, useState } from 'react';
-import isEqual from 'lodash.isequal';
+import { processHistoryItems } from 'utils/historyProcessing';
 import HistoryRow from './HistoryRow';
 import HistoryFilter, {
   filterByType,
@@ -35,20 +34,6 @@ const StyledTable = styled.table`
 interface Props {
   query: any;
 }
-
-export const formatUserName = (historyItem: any) => {
-  const isAnalyst =
-    historyItem.sessionSub.includes('idir') || historyItem.externalAnalyst;
-
-  const fullName = isAnalyst
-    ? `${historyItem.givenName} ${historyItem.familyName}`
-    : 'The applicant';
-  const isSystem = historyItem.createdBy === 1 && historyItem.op === 'INSERT';
-  return {
-    ...historyItem,
-    user: isSystem ? 'The system' : fullName,
-  };
-};
 
 const HistoryTable: React.FC<Props> = ({ query }) => {
   const queryFragment = useFragment<HistoryTable_query$key>(
@@ -91,49 +76,30 @@ const HistoryTable: React.FC<Props> = ({ query }) => {
   const originalOrganizationName =
     formData?.jsonData?.organizationProfile?.organizationName;
 
-  const applicationHistory = useMemo(() => {
-    return [...history.nodes]?.sort((a, b) => {
-      // sort by updated at if the record was deleted
-      const aDeleted = a.op === 'UPDATE';
-      const bDeleted = b.op === 'UPDATE';
-      const aDate = aDeleted ? a.record.updated_at : a.createdAt;
-      const bDate = bDeleted ? b.record.updated_at : b.createdAt;
-      return new Date(aDate).getTime() - new Date(bDate).getTime();
+  const processedHistory = useMemo(() => {
+    return processHistoryItems(history.nodes, {
+      includeAttachments: false,
+      applyUserFormatting: true,
     });
   }, [history.nodes]);
 
-  const receivedIndex = useMemo(
-    () =>
-      applicationHistory
-        .map((historyItem) => historyItem.item)
-        .indexOf('received'),
-    [applicationHistory]
-  );
-
-  const historyList = useMemo(() => {
-    return applicationHistory
-      .slice(receivedIndex)
-      .reverse()
-      .map(formatUserName);
-  }, [applicationHistory, receivedIndex]);
-
-  const typeOptions = useMemo(
-    () => getTypeOptions(historyList, filters),
-    [historyList, filters]
-  );
-  const userOptions = useMemo(
-    () => getUserOptions(historyList, filters),
-    [historyList, filters]
-  );
-
-  const filteredHistory = useMemo(
-    () =>
-      historyList.filter(
+  const filteredHistory = useMemo(() => {
+    return processedHistory
+      .map(({ historyItem }) => historyItem)
+      .filter(
         (historyItem) =>
           filterByType(historyItem, filters) &&
           filterByUser(historyItem, filters)
-      ),
-    [historyList, filters]
+      );
+  }, [processedHistory, filters]);
+
+  const typeOptions = useMemo(
+    () => getTypeOptions(filteredHistory, filters),
+    [filteredHistory, filters]
+  );
+  const userOptions = useMemo(
+    () => getUserOptions(filteredHistory, filters),
+    [filteredHistory, filters]
   );
 
   const chronologicalFilteredHistory = filteredHistory.slice().reverse();
@@ -175,110 +141,33 @@ const HistoryTable: React.FC<Props> = ({ query }) => {
 
       <StyledTable cellSpacing="0" cellPadding="0">
         <tbody>
-          {filteredHistory?.map((historyItem, index, array) => {
-            const { recordId } = historyItem;
-            const a = array.slice(index + 1);
-            let prevItems;
-            if (historyItem.tableName === 'attachment') {
-              totalLength--;
-            }
-            if (historyItem.op === 'UPDATE') {
-              prevItems = [{ record: historyItem.oldRecord }];
-            } else {
-              prevItems = a.filter((previousItem) => {
-                // assessment data must match by item type
-                if (previousItem.tableName === 'assessment_data') {
-                  return (
-                    previousItem.tableName === historyItem.tableName &&
-                    previousItem.item === historyItem.item
-                  );
-                }
-                // rfis must match by rfi_number
-                if (previousItem.tableName === 'rfi_data') {
-                  return (
-                    previousItem.tableName === historyItem.tableName &&
-                    previousItem.record.rfi_number ===
-                      historyItem.record.rfi_number &&
-                    previousItem.op === 'INSERT'
-                  );
-                }
-                // community reports must match by quarter
-                if (
-                  previousItem.tableName ===
-                    'application_community_progress_report_data' &&
-                  previousItem.tableName === historyItem.tableName
-                ) {
-                  const quarter =
-                    historyItem.record.json_data.dueDate &&
-                    getFiscalQuarter(historyItem.record.json_data.dueDate);
-                  const year =
-                    historyItem.record.json_data.dueDate &&
-                    getFiscalYear(historyItem.record.json_data.dueDate);
-                  const updated =
-                    previousItem.op === 'INSERT' &&
-                    getFiscalQuarter(previousItem.record.json_data.dueDate) ===
-                      quarter &&
-                    getFiscalYear(previousItem.record.json_data.dueDate) ===
-                      year;
-                  return updated;
-                }
-                // application milestone needs to match by quarter
-                if (
-                  previousItem.tableName === 'application_milestone_data' &&
-                  previousItem.tableName === historyItem.tableName
-                ) {
-                  const quarter =
-                    historyItem.record.json_data.dueDate &&
-                    getFiscalQuarter(historyItem.record.json_data.dueDate);
-                  const year =
-                    historyItem.record.json_data.dueDate &&
-                    getFiscalYear(historyItem.record.json_data.dueDate);
-                  const updated =
-                    previousItem.op === 'INSERT' &&
-                    getFiscalQuarter(previousItem.record.json_data.dueDate) ===
-                      quarter &&
-                    getFiscalYear(previousItem.record.json_data.dueDate) ===
-                      year;
-                  return updated;
-                }
-                // change request data must match by amendment number
-                if (previousItem.tableName === 'change_request_data') {
-                  return (
-                    previousItem.tableName === historyItem.tableName &&
-                    previousItem.record.json_data.amendmentNumber ===
-                      historyItem.record.json_data.amendmentNumber
-                  );
-                }
-                return previousItem.tableName === historyItem.tableName;
-              });
-            }
+          {processedHistory
+            .filter(
+              ({ historyItem }) =>
+                filterByType(historyItem, filters) &&
+                filterByUser(historyItem, filters)
+            )
+            .map(({ historyItem, prevHistoryItem }, index) => {
+              const { recordId } = historyItem;
 
-            const prevHistoryItem = prevItems.length > 0 ? prevItems[0] : {};
-            // Skipping duplicate history items for communities
-            if (
-              historyItem?.tableName === 'application_communities' &&
-              isEqual(
-                prevHistoryItem.record?.application_rd,
-                historyItem.record?.application_rd
-              )
-            ) {
-              totalLength--;
-              return null;
-            }
-            // using index + recordId for key as just recordId was causing strange duplicate record bug for delete history item until page refresh
-            return (
-              <HistoryRow
-                // eslint-disable-next-line react/no-array-index-key
-                key={index + recordId}
-                historyItem={historyItem}
-                prevHistoryItem={prevHistoryItem}
-                originalProjectTitle={originalProjectTitle}
-                originalOrganizationName={originalOrganizationName}
-                recordWithOrgChange={recordWithOrgChange}
-                recordWithTitleChange={recordWithTitleChange}
-              />
-            );
-          })}
+              if (historyItem.tableName === 'attachment') {
+                totalLength--;
+              }
+
+              // using index + recordId for key as just recordId was causing strange duplicate record bug for delete history item until page refresh
+              return (
+                <HistoryRow
+                  // eslint-disable-next-line react/no-array-index-key
+                  key={index + recordId}
+                  historyItem={historyItem}
+                  prevHistoryItem={prevHistoryItem}
+                  originalProjectTitle={originalProjectTitle}
+                  originalOrganizationName={originalOrganizationName}
+                  recordWithOrgChange={recordWithOrgChange}
+                  recordWithTitleChange={recordWithTitleChange}
+                />
+              );
+            })}
         </tbody>
       </StyledTable>
       <div>TOTAL LENGTH: {totalLength}</div>
